new tool sc./enemy weapon.hurt timer fix
added a utterly useless tool that stops the anim_tree when in editor. Supposedly this will help with the git commits, especially the enemy and player scenes, to be less crowded... (the bone pos are updated in scene too.)
To test if the enemy feels/acts good enough in prototype I manually coded in their timers of "hurting". I can't and don't need to think of a new way that would make me able to know when someones attacking.
Although I could seperate the main attack animations to 2 parts now, initial_attack and damaging_attack (or something like that), that would help me as I could check the current anim and based on that set the "hurting" to true